Move login page container store to central auth store
