--- src/containers/LoginPage/constants.js
@@ -1,5 +1,0 @@
-export const LOGIN = '[Login]';
-
-export const LOGIN_REQUEST = `${LOGIN} LOGIN_REQUEST`;
-export const LOGIN_SUCCESS = `${LOGIN} LOGIN_SUCCESS`;
-export const LOGIN_ERROR = `${LOGIN} LOGIN_ERROR`;
