+++ src/containers/LoginPage/index.js
@@ -3,29 +3,20 @@ import { useDispatch, useSelector } from 'react-redux';
 import { Helmet } from 'react-helmet-async';
 import { useIntl } from 'react-intl';
 import { Link } from 'react-router-dom';
-import { useInjectReducer } from 'utils/injectReducer';
-import { useInjectSaga } from 'utils/injectSaga';
 import { makeSelectIsSocialAuthPending } from 'containers/SocialAuth/selectors';
-import { makeSelectIsLoginPending } from './selectors';
-import { login } from './actions';
+import { makeSelectIsLoginPending } from 'store/auth/selectors';
+import { login } from 'store/auth/actions';
 import SocialAuth from 'containers/SocialAuth';
 import LoginForm from './LoginForm';
-import reducer from './reducer';
-import saga from './saga';
 import messages from './messages';
 import { REGISTER, FORGOT_PASSWORD } from 'routes';
 
-const key = 'login';
-
 function LoginPage() {
-  useInjectReducer({ key, reducer });
-  useInjectSaga({ key, saga });
-
   const dispatch = useDispatch();
   const isLoginPending = useSelector(makeSelectIsLoginPending());
   const isSocialAuthPending = useSelector(makeSelectIsSocialAuthPending());
   const submitLoginForm = useCallback((...args) => dispatch(login(...args)), [
-    dispatch
+    dispatch,
   ]);
 
   const { formatMessage } = useIntl();
