+++ src/store/auth/actionTypes.js
@@ -1,5 +1,8 @@
 export const AUTH = '[Auth]';
 
+export const LOGIN_REQUEST = `${AUTH} LOGIN_REQUEST`;
+export const LOGIN_SUCCESS = `${AUTH} LOGIN_SUCCESS`;
+export const LOGIN_ERROR = `${AUTH} LOGIN_ERROR`;
 export const FETCH_AUTHENTICATED_USER_REQUEST = `${AUTH} FETCH_AUTHENTICATED_USER_REQUEST`;
 export const FETCH_AUTHENTICATED_USER_SUCCESS = `${AUTH} FETCH_AUTHENTICATED_USER_SUCCESS`;
 export const LOGOUT_REQUEST = `${AUTH} LOGOUT_REQUEST`;
