+++ src/store/auth/reducer.js
@@ -1,6 +1,9 @@
 import produce from 'immer';
 import { getItem } from 'utils/localStorage';
 import {
+  LOGIN_REQUEST,
+  LOGIN_SUCCESS,
+  LOGIN_ERROR,
   FETCH_AUTHENTICATED_USER_SUCCESS,
   LOGOUT_SUCCESS,
   SET_TOKEN,
@@ -8,6 +11,7 @@ import {
 } from './actionTypes';
 
 export const initialState = {
+  isPending: false,
   token: getItem('token') || null,
   user: null,
 };
@@ -16,6 +20,13 @@ export const initialState = {
 const appReducer = (state = initialState, action) =>
   produce(state, (draft) => {
     switch (action.type) {
+      case LOGIN_REQUEST:
+        draft.isPending = true;
+        break;
+      case LOGIN_SUCCESS:
+      case LOGIN_ERROR:
+        draft.isPending = false;
+        break;
       case FETCH_AUTHENTICATED_USER_SUCCESS:
         draft.user = action.user;
         break;
