+++ src/store/auth/sagas.js
@@ -1,11 +1,47 @@
 import { takeLatest, call, put } from 'redux-saga/effects';
-import { removeItem } from 'utils/localStorage';
+import { push } from 'connected-react-router';
+import { setItem, removeItem } from 'utils/localStorage';
 import request from 'utils/request';
-import { fetchAuthenticatedUserSuccess, logoutSuccess } from './actions';
 import {
+  fetchAuthenticatedUser,
+  fetchAuthenticatedUserSuccess,
+  logoutSuccess,
+  loginSuccess,
+  loginError,
+  setToken,
+} from './actions';
+import { enqueueSnackbar } from 'containers/Notifier/actions';
+import {
+  LOGIN_REQUEST,
   FETCH_AUTHENTICATED_USER_REQUEST,
   LOGOUT_REQUEST,
 } from './actionTypes';
+import { DASHBOARD } from 'routes';
+import messages from 'containers/LoginPage/messages';
+
+export function* authorize({ email, password }) {
+  try {
+    const { accessToken: token } = yield call(request, {
+      url: '/auth/login',
+      method: 'post',
+      data: { email, password },
+    });
+    yield put(loginSuccess());
+    yield call(setItem, 'token', token);
+    yield put(setToken(token));
+    yield put(fetchAuthenticatedUser());
+    yield put(push(DASHBOARD));
+  } catch (error) {
+    if (error.status === 401) {
+      yield put(
+        enqueueSnackbar({
+          message: messages.unauthorized,
+        })
+      );
+    }
+    yield put(loginError());
+  }
+}
 
 export function* fetchUser() {
   try {
@@ -33,6 +69,7 @@ export function* logout() {
 }
 
 export default function* appSaga() {
+  yield takeLatest(LOGIN_REQUEST, authorize);
   yield takeLatest(FETCH_AUTHENTICATED_USER_REQUEST, fetchUser);
   yield takeLatest(LOGOUT_REQUEST, logout);
 }
